+++ core/src/main/java/com/orientechnologies/orient/core/serialization/serializer/record/binary/ORecordSerializerBinary.java
@@ -40,7 +40,7 @@ public class ORecordSerializerBinary implements ORecordSerializer {
   public byte[] toStream(ORecordInternal<?> iSource, boolean iOnlyDelta) {
     checkTypeODocument(iSource);
     BytesContainer container = new BytesContainer();
-    short pos = container.alloc((short) 1);
+    int pos = container.alloc(1);
     container.bytes[pos] = LAST_RECORD_VERSION;
     serializerByVersion[LAST_RECORD_VERSION].serialize((ODocument) iSource, container);
     return container.bytes;
